+++ sdk/scvmm/azure-resourcemanager-scvmm/src/main/java/com/azure/resourcemanager/scvmm/models/GuestAgentProperties.java
@@ -51,6 +51,11 @@ public final class GuestAgentProperties implements JsonSerializable<GuestAgentPr
      */
     private ProvisioningState provisioningState;
 
+    /*
+     * The resource id of the private link scope this machine is assigned to, if any.
+     */
+    private String privateLinkScopeResourceId;
+
     /**
      * Creates an instance of GuestAgentProperties class.
      */
@@ -154,6 +159,28 @@ public final class GuestAgentProperties implements JsonSerializable<GuestAgentPr
     }
 
     /**
+     * Get the privateLinkScopeResourceId property: The resource id of the private link scope this machine is assigned
+     * to, if any.
+     * 
+     * @return the privateLinkScopeResourceId value.
+     */
+    public String privateLinkScopeResourceId() {
+        return this.privateLinkScopeResourceId;
+    }
+
+    /**
+     * Set the privateLinkScopeResourceId property: The resource id of the private link scope this machine is assigned
+     * to, if any.
+     * 
+     * @param privateLinkScopeResourceId the privateLinkScopeResourceId value to set.
+     * @return the GuestAgentProperties object itself.
+     */
+    public GuestAgentProperties withPrivateLinkScopeResourceId(String privateLinkScopeResourceId) {
+        this.privateLinkScopeResourceId = privateLinkScopeResourceId;
+        return this;
+    }
+
+    /**
      * Validates the instance.
      * 
      * @throws IllegalArgumentException thrown if the instance is not valid.
@@ -177,6 +204,7 @@ public final class GuestAgentProperties implements JsonSerializable<GuestAgentPr
         jsonWriter.writeJsonField("httpProxyConfig", this.httpProxyConfig);
         jsonWriter.writeStringField("provisioningAction",
             this.provisioningAction == null ? null : this.provisioningAction.toString());
+        jsonWriter.writeStringField("privateLinkScopeResourceId", this.privateLinkScopeResourceId);
         return jsonWriter.writeEndObject();
     }
 
@@ -211,6 +239,8 @@ public final class GuestAgentProperties implements JsonSerializable<GuestAgentPr
                 } else if ("provisioningState".equals(fieldName)) {
                     deserializedGuestAgentProperties.provisioningState
                         = ProvisioningState.fromString(reader.getString());
+                } else if ("privateLinkScopeResourceId".equals(fieldName)) {
+                    deserializedGuestAgentProperties.privateLinkScopeResourceId = reader.getString();
                 } else {
                     reader.skipChildren();
                 }