+++ sdk/scvmm/azure-resourcemanager-scvmm/src/main/java/com/azure/resourcemanager/scvmm/models/VirtualMachineInventoryItem.java
@@ -42,6 +42,11 @@ public final class VirtualMachineInventoryItem extends InventoryItemProperties {
     private String powerState;
 
     /*
+     * Gets vm generation.
+     */
+    private Long generation;
+
+    /*
      * Gets or sets the nic ip addresses.
      */
     private List<String> ipAddresses;
@@ -60,26 +65,6 @@ public final class VirtualMachineInventoryItem extends InventoryItemProperties {
      * Gets the tracked resource id corresponding to the inventory resource.
      */
     private String managedMachineResourceId;
-
-    /*
-     * Gets the tracked resource id corresponding to the inventory resource.
-     */
-    private String managedResourceId;
-
-    /*
-     * Gets the UUID (which is assigned by Vmm) for the inventory item.
-     */
-    private String uuid;
-
-    /*
-     * Gets the Managed Object name in Vmm for the inventory item.
-     */
-    private String inventoryItemName;
-
-    /*
-     * Provisioning state of the resource.
-     */
-    private ProvisioningState provisioningState;
 
     /**
      * Creates an instance of VirtualMachineInventoryItem class.
@@ -131,6 +116,15 @@ public final class VirtualMachineInventoryItem extends InventoryItemProperties {
      */
     public String powerState() {
         return this.powerState;
+    }
+
+    /**
+     * Get the generation property: Gets vm generation.
+     * 
+     * @return the generation value.
+     */
+    public Long generation() {
+        return this.generation;
     }
 
     /**
@@ -192,53 +186,12 @@ public final class VirtualMachineInventoryItem extends InventoryItemProperties {
     }
 
     /**
-     * Get the managedResourceId property: Gets the tracked resource id corresponding to the inventory resource.
-     * 
-     * @return the managedResourceId value.
-     */
-    @Override
-    public String managedResourceId() {
-        return this.managedResourceId;
-    }
-
-    /**
-     * Get the uuid property: Gets the UUID (which is assigned by Vmm) for the inventory item.
-     * 
-     * @return the uuid value.
-     */
-    @Override
-    public String uuid() {
-        return this.uuid;
-    }
-
-    /**
-     * Get the inventoryItemName property: Gets the Managed Object name in Vmm for the inventory item.
-     * 
-     * @return the inventoryItemName value.
-     */
-    @Override
-    public String inventoryItemName() {
-        return this.inventoryItemName;
-    }
-
-    /**
-     * Get the provisioningState property: Provisioning state of the resource.
-     * 
-     * @return the provisioningState value.
-     */
-    @Override
-    public ProvisioningState provisioningState() {
-        return this.provisioningState;
-    }
-
-    /**
      * Validates the instance.
      * 
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     @Override
     public void validate() {
-        super.validate();
         if (cloud() != null) {
             cloud().validate();
         }
@@ -272,14 +225,14 @@ public final class VirtualMachineInventoryItem extends InventoryItemProperties {
                 reader.nextToken();
 
                 if ("managedResourceId".equals(fieldName)) {
-                    deserializedVirtualMachineInventoryItem.managedResourceId = reader.getString();
+                    deserializedVirtualMachineInventoryItem.withManagedResourceId(reader.getString());
                 } else if ("uuid".equals(fieldName)) {
-                    deserializedVirtualMachineInventoryItem.uuid = reader.getString();
+                    deserializedVirtualMachineInventoryItem.withUuid(reader.getString());
                 } else if ("inventoryItemName".equals(fieldName)) {
-                    deserializedVirtualMachineInventoryItem.inventoryItemName = reader.getString();
+                    deserializedVirtualMachineInventoryItem.withInventoryItemName(reader.getString());
                 } else if ("provisioningState".equals(fieldName)) {
-                    deserializedVirtualMachineInventoryItem.provisioningState
-                        = ProvisioningState.fromString(reader.getString());
+                    deserializedVirtualMachineInventoryItem
+                        .withProvisioningState(ProvisioningState.fromString(reader.getString()));
                 } else if ("inventoryType".equals(fieldName)) {
                     deserializedVirtualMachineInventoryItem.inventoryType
                         = InventoryType.fromString(reader.getString());
@@ -291,6 +244,8 @@ public final class VirtualMachineInventoryItem extends InventoryItemProperties {
                     deserializedVirtualMachineInventoryItem.osVersion = reader.getString();
                 } else if ("powerState".equals(fieldName)) {
                     deserializedVirtualMachineInventoryItem.powerState = reader.getString();
+                } else if ("generation".equals(fieldName)) {
+                    deserializedVirtualMachineInventoryItem.generation = reader.getNullable(JsonReader::getLong);
                 } else if ("ipAddresses".equals(fieldName)) {
                     List<String> ipAddresses = reader.readArray(reader1 -> reader1.getString());
                     deserializedVirtualMachineInventoryItem.ipAddresses = ipAddresses;